refactor(@angular-cli): get rid of ast-tools and base-href-webpack
Those 2 packages were still using the wrong scope (@angular-cli) and were not used by anyone outside the CLI. Just moving the code in the main package is enough.
