--- packages/@angular-cli/ast-tools/package.json
@@ -1,32 +1,0 @@
-{
-  "name": "@angular-cli/ast-tools",
-  "version": "1.0.16",
-  "description": "CLI tool for Angular",
-  "main": "./src/index.js",
-  "keywords": [
-    "angular",
-    "cli",
-    "ast",
-    "tool"
-  ],
-  "repository": {
-    "type": "git",
-    "url": "https://github.com/angular/angular-cli.git"
-  },
-  "author": "angular",
-  "license": "MIT",
-  "bugs": {
-    "url": "https://github.com/angular/angular-cli/issues"
-  },
-  "homepage": "https://github.com/angular/angular-cli",
-  "engines": {
-    "node": ">= 4.1.0",
-    "npm": ">= 3.0.0"
-  },
-  "dependencies": {
-    "@angular/tsc-wrapped": "^0.5.0",
-    "rxjs": "^5.0.1",
-    "denodeify": "^1.2.1",
-    "typescript": "~2.0.3"
-  }
-}
