--- packages/@angular-cli/ast-tools/tsconfig.json
@@ -1,24 +1,0 @@
-{
-  "compilerOptions": {
-    "declaration": true,
-    "experimentalDecorators": true,
-    "mapRoot": "",
-    "module": "commonjs",
-    "moduleResolution": "node",
-    "noEmitOnError": true,
-    "noImplicitAny": true,
-    "outDir": "../../../dist/@angular-cli/ast-tools",
-    "rootDir": ".",
-    "sourceMap": true,
-    "sourceRoot": "/",
-    "target": "es5",
-    "lib": ["es6"],
-    "typeRoots": [
-      "../../node_modules/@types"
-    ],
-    "types": [
-      "jasmine",
-      "node"
-    ]
-  }
-}
