--- packages/@angular-cli/base-href-webpack/package.json
@@ -1,25 +1,0 @@
-{
-  "name": "@angular-cli/base-href-webpack",
-  "version": "1.0.16",
-  "description": "Base HREF Webpack plugin",
-  "main": "./src/index.js",
-  "keywords": [
-    "angular",
-    "cli",
-    "webpack",
-    "plugin",
-    "tool"
-  ],
-  "repository": {
-    "type": "git",
-    "url": "https://github.com/angular/angular-cli.git"
-  },
-  "author": "angular",
-  "license": "MIT",
-  "bugs": {
-    "url": "https://github.com/angular/angular-cli/issues"
-  },
-  "homepage": "https://github.com/angular/angular-cli",
-  "dependencies": {
-  }
-}
