--- packages/@angular-cli/base-href-webpack/tsconfig.json
@@ -1,27 +1,0 @@
-{
-  "compilerOptions": {
-    "declaration": true,
-    "experimentalDecorators": true,
-    "mapRoot": "",
-    "module": "commonjs",
-    "moduleResolution": "node",
-    "noEmitOnError": true,
-    "noImplicitAny": true,
-    "outDir": "../../../dist/@angular-cli/base-href-webpack",
-    "rootDir": ".",
-    "sourceMap": true,
-    "sourceRoot": "/",
-    "target": "es5",
-    "lib": ["es6"],
-    "typeRoots": [
-      "../../node_modules/@types"
-    ],
-    "types": [
-      "jasmine",
-      "node"
-    ]
-  },
-  "exclude": [
-    "**/*.spec.ts"
-  ]
-}
