+++ packages/@angular/cli/blueprints/pipe/index.ts
@@ -1,3 +1,5 @@
+import {NodeHost} from '../../lib/ast-tools';
+
 const path = require('path');
 const fs = require('fs');
 const chalk = require('chalk');
@@ -5,7 +7,6 @@ const dynamicPathParser = require('../../utilities/dynamic-path-parser');
 const stringUtils = require('ember-cli-string-utils');
 const astUtils = require('../../utilities/ast-utils');
 const findParentModule = require('../../utilities/find-parent-module').default;
-const NodeHost = require('@angular-cli/ast-tools').NodeHost;
 const Blueprint = require('../../ember-cli/lib/models/blueprint');
 const getFiles = Blueprint.prototype.files;
