+++ packages/@angular/cli/lib/ast-tools/index.ts
