+++ packages/@angular/cli/lib/ast-tools/route-utils.ts
