+++ packages/@angular/cli/lib/ast-tools/spec-utils.ts
