+++ packages/@angular/cli/lib/base-href-webpack/base-href-webpack-plugin.ts
