+++ packages/@angular/cli/lib/base-href-webpack/index.ts
