+++ packages/@angular/cli/package.json
@@ -26,11 +26,10 @@
   },
   "homepage": "https://github.com/angular/angular-cli",
   "dependencies": {
-    "@angular-cli/ast-tools": "^1.0.14",
-    "@angular-cli/base-href-webpack": "^1.0.14",
     "@angular/compiler": "^2.3.1",
     "@angular/compiler-cli": "^2.3.1",
     "@angular/core": "^2.3.1",
+    "@angular/tsc-wrapped": "^0.5.0",
     "@ngtools/json-schema": "^1.0.0",
     "@ngtools/webpack": "^1.2.3",
     "async": "^2.1.4",
@@ -74,6 +73,7 @@
     "resolve": "^1.1.7",
     "rimraf": "^2.5.3",
     "rsvp": "^3.0.17",
+    "rxjs": "^5.0.1",
     "sass-loader": "^4.1.1",
     "script-loader": "^0.7.0",
     "semver": "^5.1.0",
