+++ packages/@angular/cli/tsconfig.json
@@ -20,9 +20,6 @@
     ],
     "baseUrl": "",
     "paths": {
-      "@angular-cli/ast-tools": [ "../../../dist/@angular-cli/ast-tools/src" ],
-      "@angular-cli/base-href-webpack": [ "../../../dist/@angular-cli/base-href-webpack/src" ],
-      "@angular-cli/version": [ "../../../dist/@angular-cli/version/src" ],
       "@ngtools/json-schema": [ "../../../dist/@ngtools/json-schema/src" ],
       "@ngtools/webpack": [ "../../../dist/@ngtools/webpack/src" ]
     }
